backend/task_manager/requirements/requirements.dev.txt
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9